\name{ACSN_Enrichment-package}
\alias{ACSN_Enrichment-package}
\alias{ACSN_Enrichment}
\docType{package}
\title{
\packageTitle{ACSN_Enrichment}
}
\description{
\packageDescription{ACSN_Enrichment}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ACSN_Enrichment}
\packageIndices{ACSN_Enrichment}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ACSN_Enrichment}

Maintainer: \packageMaintainer{ACSN_Enrichment}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
