% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\docType{data}
\name{ACSN_maps}
\alias{ACSN_maps}
\title{Atlas of Cancer Signalling Networks}
\format{A list of dataframes
\describe{
 \item{Apoptosis}{Map of apoptosis pathways}
 \item{CellCycle}{Map of the cell cycle pathways}
 \item{DNA_repair}{Map of DNA repair}
 \item{EMT_motility}{Map of the Epithelial Mesenchymal Transition}
 \item{Master}{Map grouping all modules from other maps, without a master module for each map}
 \item{Survival}{Map of cellular survival pathways}
}}
\source{
\url{https://acsn.curie.fr/downloads.html}
}
\usage{
ACSN_maps
}
\description{
A dataset containing the six maps of ACSN: apoptosis, cell cycle, DNA reparation, EMT motility, survival, and the master map
}
\keyword{datasets}

