% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{Create_master_map}
\alias{Create_master_map}
\title{From a list of maps, create  or replace a master}
\usage{
Create_master_map(maps)
}
\arguments{
\item{maps}{A list of molecular maps created by `format_from_gmt`}
}
\value{
Returns a list with previous maps and the master map, i.e. a concatenation of previous maps.
}
\description{
From a list of maps, create  or replace a master
}
\examples{
Create_master_map(list(Cycle = ACSNMineR::ACSN_maps$CellCycle,
                   Apoptosis = ACSNMineR::ACSN_maps$Apoptosis))
}

