% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{format_from_gmt}
\alias{format_from_gmt}
\title{Pre-processing data}
\usage{
format_from_gmt(path = "")
}
\arguments{
\item{path}{Path to the gmt file to be imported}
}
\description{
Pre-processing data
}

