% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\docType{data}
\name{genes_test}
\alias{genes_test}
\title{Set of genes to test map}
\format{A character vector}
\usage{
genes_test
}
\description{
Genes of high importance in oncogenesis
}
\keyword{datasets}

