% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{multisample_enrichment}
\alias{multisample_enrichment}
\title{Enrichment analysis function}
\usage{
multisample_enrichment(Genes_by_sample = NULL,
  maps = list(ACSN_apop_formatted, ACSN_cellcyc_formatted,
  ACSN_dnarep_formatted, ACSN_emtmotil_formatted, ACSN_master_formatted,
  ACSN_survival_formatted), correction_multitest = FALSE,
  statistical_test = "fisher", min_module_size = 5, universe = "HUGO",
  individual_threshold = 1, cohort_threshold = 1)
}
\arguments{
\item{Genes_by_sample}{List of character vectors. Each list element name should be a sample name, and each character vector the set of genes to test for the sample.}

\item{maps}{list of maps generated by format_from_gmt. Default: tests on all acsn maps}

\item{correction_multitest}{either FALSE, "bonferroni", "holm", "hochberg", "hommel", "BH", "fdr" (identical to BH), or "BY"}

\item{min_module_size}{will remove from the analysis all modules which are (strictly) smaller than threshold}

\item{universe}{Universe on which the statistical analysis should be performed. Can be either "HUGO","ACSN", or a character vector of genes.}

\item{statistical}{test one of "fisher", "hypergeom"}

\item{threshold}{maximal p-value (corrected if correction is enabled) that will be displayed}

\item{cohort_threshold:}{if TRUE modules will be kept in all samples if at least one sample has p-value lower than threshold, otherwise the threshold is applied for each sample independently.}
}
\description{
Allows gene set enrichment analysis
}

