% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{p.val.calc}
\alias{p.val.calc}
\title{Calculate p-value}
\usage{
p.val.calc(x, y, z, a, stat_test, alt)
}
\arguments{
\item{x}{: first value}

\item{y}{: second value}

\item{z}{: third value}

\item{a}{: fourth value}

\item{stat_test}{: statistical test to be used}

\item{alt}{: alternative: one of two-sided, greater, less or both}
}
\description{
Calculate p-value
}

