% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Scale for barplots and heatmaps}
\usage{
reverselog_trans(base = 10)
}
\arguments{
\item{base}{: base for the log, defaut is e}
}
\description{
Outputs the "-log" of a scale
}

