% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Gene set enrichment analysis}
\usage{
enrichment(Genes = NULL, maps = ACSNEnrichment::ACSN_maps,
  correction_multitest = "BH", statistical_test = "fisher",
  min_module_size = 5, universe = "map_defined", threshold = 0.05)
}
\arguments{
\item{Genes}{Character vector of genes that should be tested for enrichment}

\item{maps}{list of maps generated by format_from_gmt. Names of element of list will be used to track modules.
Default: tests on all acsn maps.}

\item{correction_multitest}{either FALSE, "bonferroni", "holm", "hochberg", "hommel", "BH", "fdr" (identical to BH), or "BY"}

\item{statistical_test}{one of "fisher", "hypergeom"}

\item{min_module_size}{will remove from the analysis all modules which are (strictly) smaller than threshold}

\item{universe}{Universe on which the statistical analysis should be performed.
Can be either "HUGO","ACSN","map_defined", or a character vector of genes.}

\item{threshold}{: maximal p-value (corrected if correction is enabled) that will be displayed}
}
\description{
Compute and represent gene set enrichment from your data based on pre-saved maps from ACSN or user imported maps.
The gene set enrichment can be run with hypergeometric test or Fisher exact test,
and can use multiple corrections. Visualization of data can be done either by barplots or heatmaps.
}
\examples{
enrichment(genes_test,min_module_size = 10,
   threshold = 0.05,
   maps = list(cellcycle = ACSNEnrichment::ACSN_maps$CellCycle))
}

