% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{format_from_gmt}
\alias{format_from_gmt}
\title{Import data from gmt files
Convert gmt file to dataframe that can be used for anaysis}
\usage{
format_from_gmt(path = "")
}
\arguments{
\item{path}{Path to the gmt file to be imported}
}
\description{
Import data from gmt files
Convert gmt file to dataframe that can be used for anaysis
}
\examples{
file<-system.file("extdata", "cellcycle_short.gmt", package = "ACSNMineR")
format_from_gmt(file)
}

